moved the components about
